Move template components to own folder
